at machine Dell at network NetVirtua 10Mbps at Santos, SP - Brazil, at laptop Dell XPS L502X processor i7 2670QM 8GB RAM 1 TB Disk 1.5TB Backup Samsung Disk at Santos SP, Brazil
-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Léia Maheli de Souza Tsumanuma!! Mummi
  Tack Simon Petrus Lundgren!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabrielli Cecyn Lundgren!! Gabi
  Tack Rafael Cecyn Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
}

Obrigado Cits
{
   Obrigado Rejeane!!
   Obrigado Monique!!
   Obrigado Gilmar Tuchinski!!
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Buhrer!!
}

Obrigado Brasil
{
   Obrigado guardas!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado moças e mulheres...
   Obrigado moços e homens...
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Dilma Souza!!
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}

Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt & Edras Pacola!! 
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you Anderson Silva!!
   Thank you Pietro!!
   Thank you Thiago Ferreira!! - de 1996 do grupo do Cleiton, Anderson Silva e Pietro
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
T Obrigado Telminho, você é legal!!
1 Obrigado Telminho, você é bom!!
2 Obrigado Telminho, eu gosto de você!!
n Obrigado Telminho, eu gosto muito de você!!
n Obrigado Telminho, você é demais!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a desobrigado.
+∞

+++ appseed/c/bsdiff/bsdiff.cpp
@@ -138,7 +138,7 @@ static file_offset search(file_offset *I,u_char *old,file_offset oldsize,
    };
 
    x=st+(en-st)/2;
-   if(memcmp_dup(old+I[x],_new,min(oldsize-I[x],newsize))<0) {
+   if(memcmp_dup(old+I[x],_new, (size_t) min(oldsize-I[x],newsize))<0) {
       return search(I,old,oldsize,_new,newsize,x,en,pos);
    } else {
       return search(I,old,oldsize,_new,newsize,st,x,pos);
@@ -203,7 +203,7 @@ int32_t bsdiff(const char * oldfile, const char * newfile, const char * patchfil
    //fread_dup in chunks, don't rely on fread_dup always returns full data!
    if(((fd=fopen_dup(oldfile,"rb")) ==0) ||
       ((oldsize=fseek_dup(fd,0,SEEK_END))==-1) ||
-      ((old=(u_char*)_ca_alloc(oldsize+1))==NULL) ||
+      ((old=(u_char*)_ca_alloc((size_t) (oldsize + 1)))==NULL) ||
       (fseek_dup(fd,0,SEEK_SET)!=0))
    {
       if(fd != 0)
@@ -219,7 +219,7 @@ int32_t bsdiff(const char * oldfile, const char * newfile, const char * patchfil
 
    file_offset r = oldsize;
 
-   while(r > 0 && (i = (file_offset) fread_dup(old + oldsize - r, 1, r, fd)) > 0)
+   while(r > 0 && (i = (file_offset) fread_dup(old + oldsize - r, 1, (size_t) r, fd)) > 0)
       r-=i;
    if (r>0 || fclose_dup(fd)>0)
    {
@@ -228,8 +228,8 @@ int32_t bsdiff(const char * oldfile, const char * newfile, const char * patchfil
    }
 
 
-   if(((I=(file_offset*)_ca_alloc((oldsize+1)*sizeof(file_offset)))==NULL) ||
-      ((V=(file_offset*)_ca_alloc((oldsize+1)*sizeof(file_offset)))==NULL))
+   if(((I=(file_offset*)_ca_alloc((size_t) ((oldsize+1)*sizeof(file_offset))))==NULL) ||
+      ((V=(file_offset*)_ca_alloc((size_t) ((oldsize+1)*sizeof(file_offset))))==NULL))
    {
       _ca_free(old, 0);
       if(I != NULL)
@@ -260,7 +260,7 @@ int32_t bsdiff(const char * oldfile, const char * newfile, const char * patchfil
    //fread_dup in chunks, don't rely on fread_dup always returns full data!
    if(((fd=fopen_dup(newfile,"rb"))== NULL) ||
       ((newsize=fseek_dup(fd,0,SEEK_END))==-1) ||
-      ((_new=(u_char*)_ca_alloc(newsize+1))==NULL) ||
+      ((_new=(u_char*)_ca_alloc((size_t) (newsize + 1)))==NULL) ||
       (fseek_dup(fd,0,SEEK_SET)!=0))
    {
       if(fd != 0)
@@ -273,7 +273,7 @@ int32_t bsdiff(const char * oldfile, const char * newfile, const char * patchfil
    }
 
    r=newsize;
-   while(r > 0 && (i = (file_offset) fread_dup(_new + newsize - r, 1, r, fd)) > 0)
+   while(r > 0 && (i = (file_offset) fread_dup(_new + newsize - r, 1, (size_t) r, fd)) > 0)
       r-=i;
    if (r>0 || fclose_dup(fd)> 0)
    {
@@ -283,8 +283,8 @@ int32_t bsdiff(const char * oldfile, const char * newfile, const char * patchfil
       return err(1,"%s",oldfile);
    }
 
-   if(((db=(u_char*)_ca_alloc(newsize+1))==NULL) ||
-      ((eb=(u_char*)_ca_alloc(newsize+1))==NULL))
+   if(((db=(u_char*)_ca_alloc((size_t) (newsize + 1)))==NULL) ||
+      ((eb=(u_char*)_ca_alloc((size_t) (newsize + 1)))==NULL))
    {
       _ca_free(old, 0);
       _ca_free(_new, 0);
