+++ appseed/ca2/html/html_image.cpp
@@ -17,7 +17,8 @@ namespace html
 
             pdata->m_pdc->set_alpha_mode(::ca::alpha_mode_blend);
 
-            pdata->m_pdc->BitBlt(get_x(), get_y(), get_cx(), get_cy(), pdata->m_imagea[m_iImage].m_spdib->get_graphics(), 0, 0, SRCCOPY);
+            pdata->m_pdc->BitBlt((int32_t) get_x(), (int32_t) get_y(), (int32_t) get_cx(), (int32_t) get_cy(), 
+                                 pdata->m_imagea[m_iImage].m_spdib->get_graphics(), 0, 0, SRCCOPY);
 
          }
 
@@ -31,8 +32,8 @@ namespace html
          {
             m_iImage = pdata->get_image_index(pelemental->m_propertyset["src"]);
             synch_lock lockImage(Sys(pdata->m_papp).get_twf());
-            m_cxMax = pdata->m_imagea[m_iImage].m_spdib->cx;
-            m_cxMin = pdata->m_imagea[m_iImage].m_spdib->cy;
+            m_cxMax = (float) pdata->m_imagea[m_iImage].m_spdib->cx;
+            m_cxMin = (float) pdata->m_imagea[m_iImage].m_spdib->cy;
          }
       }
 
@@ -44,10 +45,10 @@ namespace html
 
             if(lockImage.lock(duration::zero()))
             {
-               pdata->m_layoutstate.m_cx = pdata->m_imagea[m_iImage].m_spdib->cx;
+               pdata->m_layoutstate.m_cx = (float) pdata->m_imagea[m_iImage].m_spdib->cx;
                if(pdata->m_imagea[m_iImage].m_spdib->cy > pdata->m_layoutstate.m_cy)
                {
-                  pdata->m_layoutstate.m_cy = pdata->m_imagea[m_iImage].m_spdib->cy;
+                  pdata->m_layoutstate.m_cy = (float) pdata->m_imagea[m_iImage].m_spdib->cy;
                }
             }
             else
